Refactor code for getting OpenSim version.
